requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyarrow>=12.0.0
PyYAML>=6.0